+++ lib/selenium_fury/elements/element_helpers/generic_element_helper.rb
@@ -17,61 +17,68 @@ module SeleniumFury
   module Elements
       # Defines what we can do to all elements
       module GenericElementHelper
-        def el
-          raise "Locator at #{location} is not present" unless present?
-          @driver.find_element(location)
-        end
-
-        def list
-          (1..number_matching).map { |i| self.class.new({css: location[:css]+":nth-of-type(#{i})"}, @driver) }
+        def wd_element
+          raise "Locator at #{locator} is not present" unless present?
+          @driver.find_element(locator)
         end
 
         def present?
           # Set implicit wait to zero so it doesn't wait that time each method call
           implicit_wait = driver.manage.timeouts.implicit_wait
           driver.manage.timeouts.implicit_wait = 0
-          present = number_matching > 0
+          present = @driver.find_elements(locator).size > 0
           driver.manage.timeouts.implicit_wait = implicit_wait
           present
         end
 
-        def number_matching
-          @driver.find_elements(location).size
-        end
-
         # Raises error if not already present
         def visible?
-          el.displayed?
+          wd_element.displayed?
+        end
+
+        def accessible?
+          present? && visible?
         end
 
         def value
-          el.attribute('value')
+          wd_element.attribute('value')
         end
 
-        def move_to
-          @driver.action.move_to(el).perform
+        def move_to(right_by=nil, down_by=nil)
+          @driver.action.move_to(wd_element, right_by, down_by).perform
+          self
+        end
+
+        def drag_and_drop(target)
+          @driver.action.drag_and_drop(wd_element, target.wd_element).perform
+        end
+
+        def attach_file(path)
+          send_keys!(path)
         end
 
         def double_click
-          @driver.action.double_click(el).perform
+          @driver.action.double_click(wd_element).perform
         end
 
         def double_click!
-          el.click
+          wd_element.click
         rescue Exception => e
-          puts "Encountered #{e.class} trying to click element at #{self.location}"
+          puts "Encountered #{e.class} trying to click element at #{self.locator}"
         ensure
-          el.click
+          wd_element.click
         end
 
+        # TODO - For proper encapsulation, need to define what is and what isn't allowed for each element type
         # Use any methods from WebDriverElement not present
         def method_missing method_sym, *args
-          if el.respond_to?(method_sym)
-            el.send(method_sym, *args)
+          if wd_element.respond_to?(method_sym)
+            wd_element.send(method_sym, *args)
           else
             raise NoMethodError, "No such  method #{method_sym} for #{self.class}"
           end
         end
+
       end # GenericElementHelper
     end # Elements
 end # SeleniumFury