+++ spec/selenium_fury/page_object_spec.rb
@@ -2,33 +2,29 @@ require 'spec_helper'
 
 describe SeleniumFury::PageObject do
   let(:test_page) { TestPage.new(driver) }
-  before(:all) { start_web_driver TEST_PAGE_URL }
+  before(:all) { start_web_driver(TEST_PAGE_URL) }
   after(:all) { stop_web_driver }
   after(:each) do
     Object.instance_eval { remove_const :TestPage } if Object.const_defined? :TestPage
     load 'test_page/test_page.rb'
   end
 
-  it "has Selenium::WebDriver::Elements defined with the element method" do
-    test_page.should_not be_nil
-    test_page.textarea.should be_a(Selenium::WebDriver::Element)
+  it "PageObject provides a method to return all SeleniumFury::Elements::GenericElement defined in the class" do
+    test_page.class.elements.should have(42).elements
+    test_page.send(test_page.class.elements[0]).should be_a(SeleniumFury::Elements::GenericElement)
   end
 
-  it "should call methods on Selenium::WebDriver::Elements defined in the TestPage class" do
-    test_page.input_message.send_keys 'this is awesome'
-    test_page.input_msg_button.click
-    driver.find_element(id: 'message').text.should eql 'this is awesome'
+  it "should return a list of unexpected missing elements" do
+    test_page.missing_elements.should == {:input_hidden_element=>{:id=>"input_hidden"}}
   end
 
-  it "PageObject provides a method to return all Selenium::WebDriver::Element defined in the class" do
-    test_page.class.elements.should_not be_nil
-    test_page.class.elements.should have(40).elements
-    test_page.method(test_page.class.elements[0]).call.class.should eql Selenium::WebDriver::Element
+  it "should return a list of unexpected missing elements when provided a list of elements" do
+    test_page.missing_elements([:input_button_element, :input_msg_button_element, :input_reset_element]).should be_empty
+    test_page.missing_elements([:input_hidden_element]).should  == {:input_hidden_element=>{:id=>"input_hidden"}}
   end
 
-  it "should raise an exception when the element can't be found" do
-    TestPage.element :not_a_element, {id: 'not a element'}
-    expect { test_page.not_a_element }.to raise_exception RuntimeError, 'Could not find element not_a_element'
+  it "should show on expected page" do
+    test_page.expected_location?('spec/test_page/test_page.html').should be_false
   end
 
 end