+++ spec/spec_helper.rb
@@ -17,7 +17,6 @@
 require 'selenium_fury'
 require 'active_support/inflector'
 
-require 'test_page/test_page_custom_generator_configuration'
 require 'test_page/test_page'
 
 TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"